Save a castle's tax balance as the change it applied, so saves commute
Every change to a castle's tax balance saved the balance it produced
(TaxBalance=%d), so two changes on two zone threads could save in the
opposite order and leave the row one change behind until the next save.

The compare-and-swap now reports the amount it actually moved the balance
by after its clamps at zero and at the maximum (CastleTaxBalance.h), and
that amount is saved relatively through the castle seam
(addCastleTaxBalance: TaxBalance = TaxBalance + delta). The four save
paths -- CastleInfo's two Ex forms and the manager's two -- are one pair
on CastleInfoManager, which the withdrawal now goes through too; the
castle's own changers are private to it. The owner change's reset is an
exchange whose taken amount is saved as a debit in the same statement as
the new owner (saveCastleOwner), so a credit made just before the reset
is taken with it, one made just after stays, and the row agrees whichever
save lands first. The unused whole-row CastleInfoManager::save, which
wrote the balance absolutely, is removed with its seam method.

Between two saves the row can stand below zero (a withdrawal saved ahead
of the credit it drew on). MySQL refuses that on the old int unsigned
column (ERROR 1690, BIGINT UNSIGNED value is out of range), so TaxBalance
is a signed BIGINT, with migration 003 for existing databases, and the
load clamps it into range. What remains is a save that never lands,
which now leaves the row off by that change instead of being overwritten
by the next save; the header and the fix log say so.

R5 5152 -> 5147: the five try blocks went with the removed paths.

+++ initdb/DARKEDEN.sql
@@ -906,7 +906,7 @@ CREATE TABLE `CastleInfo` (
   `Race` smallint(5) unsigned NOT NULL DEFAULT '0',
   `ItemTaxRatio` int(10) unsigned NOT NULL DEFAULT '0',
   `EntranceFee` int(10) unsigned NOT NULL DEFAULT '0',
-  `TaxBalance` int(10) unsigned NOT NULL DEFAULT '0',
+  `TaxBalance` bigint(20) NOT NULL DEFAULT '0',
   `BonusOptionType` text NOT NULL,
   `FirstResurrectZoneID` smallint(5) unsigned NOT NULL DEFAULT '0',
   `FirstResurrectX` int(10) unsigned NOT NULL DEFAULT '0',
